[wma2wav] Shut up a warning in 64-Bit "Debug" configuration.
git-svn-id: svn://svn.code.sf.net/p/muldersoft/code/trunk/Utils/wma2wav@243 8a597620-9bd1-4498-a720-a9593e25b6ad

+++ wma2wav/Utils.cpp
@@ -280,9 +280,9 @@ bool secure_load_library(HMODULE *module, const wchar_t* fileName)
 }
 
 #ifdef _DEBUG
-int dbg_printf(wchar_t *format, ...)
+size_t dbg_printf(wchar_t *format, ...)
 {
-	int len = 0;
+	size_t len = 0;
 	va_list args;
 	va_start (args, format);
 	if(format)
